Move multi_create delayed job to a helper class
